requirements.txt
google-generativeai >= 0.7.2
